requirements.txt
pandas
requests
python-dotenv
aiohttp
numpy
sqlalchemy
psycopg2